requirements.txt
flask
python-dotenv
pycoingecko
web3
plotly
requests
numpy
transformers
torch
